+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Change MSRV to 1.62.1.
+- Ignore `Update(Actor)` if profile is not found locally.
 
 ## [2.9.0] - 2024-02-07